\name{mfnode-package}
\alias{mfnode-package}
\alias{mfnode}
\docType{package}
\title{
\packageTitle{mfnode}
}
\description{
\packageDescription{mfnode}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mfnode}
\packageIndices{mfnode}
Functions for parsing log files produced by NWP models from
 Meteo France / ECMWF / ACCORD, the so-called "NODE" files
 Parsed elements are spectral and grid-point norms, runtime information, "Jo" tables or
 "VarQC" values.
 Other functions also exist for reading and writing specific Meteo France / ACCORD
  data files, so-called "FA format"
}
\author{
\packageAuthor{mfnode}

Maintainer: \packageMaintainer{mfnode}
}

\keyword{ package }
\seealso{

}
